Ensure player study event is handled by the GM
