+++ templates/actor/actor-party-sheet.hbs
@@ -189,7 +189,7 @@
 							</div>
 							<div class="data">
 								<header>
-									<div class="name">
+									<div class="name" data-tooltip="{{this.uuid}}">
 										{{name}}
 									</div>
 									<div class="blurb">
